python/requirements.txt
blinker==1.6.2
click==8.1.4
colorama==0.4.6
exceptiongroup==1.1.2
Flask==2.3.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.1.0
pluggy==1.2.0
psycopg2==2.9.9
pytest==7.4.0
tomli==2.0.1
Werkzeug==2.3.6
